% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CepalStatR-package.R
\docType{package}
\name{CepalStatR-package}
\alias{CepalStatR}
\alias{CepalStatR-package}
\title{CepalStatR: Access to the 'CEPALSTAT API'}
\description{
Explore metadata and retrieve indicators from the statistical portal of the Economic Commission for Latin America and the Caribbean \url{https://statistics.cepal.org/portal/cepalstat/}. Provides a reproducible interface to access the 'CEPALSTAT API', organize data into tidy formats, and generate both static and interactive visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Henry-Osorto/CepalStatR}
  \item \url{https://henry-osorto.github.io/CepalStatR/}
  \item Report bugs at \url{https://github.com/Henry-Osorto/CepalStatR/issues}
}

}
\author{
\strong{Maintainer}: Henry Osorto \email{henry.osorto@unah.edu.hn} (\href{https://orcid.org/0000-0002-4334-9179}{ORCID})

Authors:
\itemize{
  \item Henry Osorto \email{henry.osorto@unah.edu.hn} (\href{https://orcid.org/0000-0002-4334-9179}{ORCID})
}

}
\keyword{internal}
