% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R, R/zzz-http.R
\name{img_to_data_uri}
\alias{img_to_data_uri}
\title{Convert an image file to a data URI}
\usage{
img_to_data_uri(path)

img_to_data_uri(path)
}
\arguments{
\item{path}{Path to the image file.}
}
\value{
A character string containing a data URI, or NULL if the file does not exist.
}
\description{
Convert an image file to a data URI

Internal helper
}
\keyword{internal}
