% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranking.ods.R
\name{ranking.ods}
\alias{ranking.ods}
\title{Title}
\usage{
ranking.ods(
  id.indicator,
  language.en = TRUE,
  save = FALSE,
  height = 5,
  width = 9,
  size.title = 10,
  title = TRUE,
  caption = TRUE,
  color = "#032B47",
  color.alc = "#36B3FF"
)
}
\arguments{
\item{color.alc}{}
}
\description{
Title
}
