% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-html.R
\name{viewer_indicators_style}
\alias{viewer_indicators_style}
\title{CSS for viewer.indicators header layout}
\usage{
viewer_indicators_style()
}
\value{
An htmltools style tag.
}
\description{
CSS for viewer.indicators header layout
}
\keyword{internal}
