% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.data.R
\name{call.data}
\alias{call.data}
\title{Call Data Indicators CEPALSTAT}
\usage{
call.data(id.indicator, language.en = TRUE, notes = FALSE)
}
\arguments{
\item{id.indicator}{Indicator ID to retrieve.}

\item{language.en}{Logical. TRUE for English, FALSE for Spanish.}

\item{notes}{Logical. If TRUE, methodological notes are added when available.}
}
\value{
A data frame with indicator values and labels.
}
\description{
call.data allows querying the CEPALSTAT API to obtain a data frame
with indicator values, disaggregations, years, countries, and other
dimensions available for the selected indicator.
}
