% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.data.R
\name{call.data}
\alias{call.data}
\title{Call Data Indicators CEPALSTAT}
\usage{
call.data(id.indicator, language.en = TRUE, notes = FALSE)
}
\arguments{
\item{id.indicator}{You must determine the ID of the indicator or variable that you want to obtain.}

\item{language.en}{If true or omitted is selected, the default language will be English. Select False to choose the Spanish language.}

\item{notes}{They are the methodological notes that an indicator may have. Select TRUE to be able to add the notes to the data frame.}
}
\description{
call.data is a function that allows you to consult the CEPALSTAT API to obtain in
a data frame with the values of the desired indicator, disaggregations, by year, countries
and others that the indicator has.
}
\examples{
data <- call.data(id.indicator = 1)

}
