% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.indicators.R
\name{call.indicators}
\alias{call.indicators}
\title{Call indicators CEPALSTAT}
\usage{
call.indicators(language.en = TRUE)
}
\arguments{
\item{language.en}{If true or omitted is selected, the default language will be English.
Select False to choose the Spanish language.}
}
\description{
It allows generating a data frame that contains each of the indicators by areas
and subdimensions of CEPALSTAT.
}
\examples{
data.indicators <- call.indicators()
data.indicators <- call.indicators(language.en = FALSE) # To obtain the indicators in Spanish language
}
