% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz-http.R
\name{cepal_get}
\alias{cepal_get}
\title{Internal helper to retrieve CEPALSTAT resources}
\usage{
cepal_get(url, format = c("json", "csv"), timeout_sec = 60)
}
\description{
Internal helper to retrieve CEPALSTAT resources
}
\keyword{internal}
