% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CepalStatR-package.R
\docType{package}
\name{CepalStatR-package}
\alias{CepalStatR}
\alias{CepalStatR-package}
\title{CepalStatR: CepalStatR: a package in R for access to ECLAC statistics}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \doi{10.18687/LACCEI2024.1.1.1473}
  \item Report bugs at \url{https://github.com/Henry-Osorto/CepalStatR/issues}
}

}
\keyword{internal}
