% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CepalStatR-package.R
\docType{package}
\name{CepalStatR-package}
\alias{CepalStatR}
\alias{CepalStatR-package}
\title{CepalStatR: Access to ECLAC (CEPALSTAT) Statistics via R}
\description{
Tools to explore metadata and retrieve indicators from CEPALSTAT Open Data/API, building tidy panels and basic visualizations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Henry-Osorto/CepalStatR}
  \item \url{https://henry-osorto.github.io/CepalStatR/}
  \item Report bugs at \url{https://github.com/Henry-Osorto/CepalStatR/issues}
}

}
\author{
\strong{Maintainer}: Henry Osorto \email{henry.osorto@unitec.edu.hn} (\href{https://orcid.org/0000-0002-4334-9179}{ORCID})

}
\keyword{internal}
