% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{countries}
\alias{countries}
\title{list of countries}
\usage{
countries(language.en = TRUE)
}
\arguments{
\item{language.en}{If TRUE or omitted is selected, the default language will be English.
Select FALSE to choose the Spanish language.}
}
\description{
A short description...
}
\examples{
data <- countries()


}
