% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramids.R
\name{pyramids}
\alias{pyramids}
\title{Population pyramids}
\usage{
pyramids(country, years, language.en, language.en = TRUE, color = c('#B20B27', '#0A1873'),
         save = FALSE, height = 5, width = 7)
}
\arguments{
\item{width}{}
}
\description{
A short description...
}
