% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyramids.R
\name{pyramids}
\alias{pyramids}
\title{Population pyramids}
\usage{
pyramids(country, years, language.en, language.en = TRUE, color = c('#B20B27', '#0A1873'),
         save = FALSE, height = 5, width = 7)
}
\arguments{
\item{country}{Write the name of the country}

\item{years}{It is a vector that contains values between 1 and 31. These values reflect the position of a vector of years in 5-year intervals.
The starting year is 1950 (1) and the ending year is 2100 (31).}

\item{language.en}{If TRUE or omitted is selected, the default language will be English.
Select FALSE to choose the Spanish language.}

\item{color}{Contains two RGB scale characters. However, you can use the conventional names (red, yellow, blue, green, etc.)}

\item{save}{It is a logical parameter that allows saving the graph in png format in the setwd('~/') directory.
By default the save option is FALSE. Select TRUE if you want to save the graph.}

\item{height}{Determines the height of the graph. You can adjust the size by increasing or decreasing.
By default the height is 5 units ("in", "cm", "mm", or "px", depending on the device).}

\item{width}{Determines the width of the graph. You can adjust the size by increasing or decreasing.
By default the width is 7 units ("in", "cm", "mm", or "px", depending on the device).}
}
\description{
A short description...
}
\examples{
pyramids(country = "Latin America", years = 16)

# With the values of the years parameter the years 1950, 2000, 2025,
# 2050 and 2100 will be obtained.
pyramids(country = "Latin America", years = c(1, 11, 16, 21, 25, 31))

# To save the graph in setwd('~/') set save to TRUE
pyramids(country = "Latin America", years = c(1, 11, 16, 21, 25, 31),
         save = TRUE, height = 8, width = 10)
}
