% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.indicators.R
\name{viewer.indicators}
\alias{viewer.indicators}
\title{Viewer indicators CEPALSTAT}
\usage{
viewer.indicators(language.en = TRUE)
}
\arguments{
\item{language.en}{If true or omitted is selected, the default language will be English. Select False to choose the Spanish language.}
}
\description{
Allows you to view each of the indicators of the CEPALSTAT database in the R viewer window.
}
\examples{
viewer.indicators()
viewer.indicators(language.en = FALSE) # To obtain the indicators in Spanish language
}
